Mail service, exception handlers and validation

+++ backend/rest/MerchantRest/src/main/java/app/services/InsuranceService.java
@@ -13,89 +13,76 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import app.commons.Consts;
 import app.model.Insurance;
-import app.repository.local.InsuranceRepository;
+import app.repository.InsuranceRepository;
 import app.services.exceptions.BadRequestException;
 import app.services.exceptions.NotFoundException;
-import app.validators.InsuranceValidation;
 
 @Service
 public class InsuranceService {
 	@Autowired
 	private InsuranceRepository insuranceRepository;
-	public static final int JMBG_OK = 0;
-	public static final int JMBG_INVALID = 1;
-	public static final int JMBG_WARNING = 2;
-	
+
 	private static Pattern jmbgPattern = Pattern.compile("\\d{13}");
 	private static SimpleDateFormat sdf = new SimpleDateFormat("ddMMyyyy");
-
-
 	private static final Logger logger = Logger.getLogger(InsuranceService.class);
 
 	public Map<String, Object> create(Insurance insurance) {
 
 		Map<String, Object> response;
-
-		if (!InsuranceValidation.validateInsurance(insurance)) {
-			throw new BadRequestException();
-		}
-
 		response = new LinkedHashMap<String, Object>();
 		insuranceRepository.save(insurance);
-		response.put("message", "Insurance created successfully");
+		response.put("message", "Insurance successfully created");
 		response.put("insurance", insurance);
 
-		logger.info("Insurance created successfully");
+		logger.info("Insurance successfully created");
 
 		return response;
 	}
 
 	public List<Insurance> findAll() {
 
-		logger.info("List all insurances");
 		return insuranceRepository.findAll();
 	}
 
 	public Insurance findById(String insId) {
-
+		
+		if (insId == null) {
+			throw new BadRequestException("insuranceId is null");
+		}
+		
 		Insurance insurance = insuranceRepository.findOne(insId);
-		logger.info("Find insurance with id: " + insId);
-
 		if (insurance == null) {
-			throw new NotFoundException();
+			throw new NotFoundException("Insurance with id " + insId + " doesn't exist.");
 		}
 
 		return insurance;
 	}
 
 	public String remove(String insId) {
-
+		
+		if (insId == null) {
+			throw new BadRequestException("insurance id is null");
+		}
+		
 		logger.info("Removing insurance with id: " + insId);
 		insuranceRepository.delete(insId);
 		return "removed";
 	}
 
 	public Map<String, Object> edit(Insurance insurance) {
-
 		Map<String, Object> response;
-
-		if (!InsuranceValidation.validateInsurance(insurance)) {
-			throw new BadRequestException();
-		}
-
 		response = new LinkedHashMap<String, Object>();
 		insuranceRepository.save(insurance);
-		response.put("message", "Insurance updated successfully");
+		response.put("message", "Insurance successfully updated");
 		response.put("insurance", insurance);
-
-		logger.info("Insurance updated successfully");
-
+		logger.info("Insurance successfully updated");
 		return response;
 	}
-	
-	//validacija jmbg
-	
+
+	// validacija jmbg
+
 	public boolean checkJmbg(String jmbg, boolean withChecksum) {
 		if (jmbg == null)
 			return false;
@@ -111,12 +98,12 @@ public class InsuranceService {
 
 	public static int checkJmbgWithWarning(String jmbg) {
 		if (jmbg == null)
-			return JMBG_INVALID;
+			return Consts.JMBG_INVALID;
 		Matcher matcher = jmbgPattern.matcher(jmbg);
 		if (!matcher.matches())
-			return JMBG_INVALID;
+			return Consts.JMBG_INVALID;
 		if (extractDateFromJmbg(jmbg) == null)
-			return JMBG_INVALID;
+			return Consts.JMBG_INVALID;
 
 		return jmbgChecksumWithWarning(jmbg);
 	}
@@ -171,9 +158,9 @@ public class InsuranceService {
 			check = 0;
 
 		if (h != 6)
-			return m == check ? JMBG_OK : JMBG_INVALID;
+			return m == check ? Consts.JMBG_OK : Consts.JMBG_INVALID;
 		else
-			return m == check ? JMBG_OK : JMBG_WARNING;
+			return m == check ? Consts.JMBG_OK : Consts.JMBG_WARNING;
 	}
 
 	private static boolean jmbgChecksum(String jmbg) {
